Add the option to set a custom theme colour.
Also include an example to demonstrate this. Document the feature in the README
too, with a big fat warning that it should normally not be used. That same
warning is present in the example demonstrating the use of a custom theme
colour. The only real usecase of this feature could possibly be to make colours
darker, in case you need to present on a bad beamer. Maybe.