Added the option tuefont and sans
The option tuefont can be used to not use the tuefont by setting it to
false, the default value is true (which uses the tuefont). The option sans
can be used if you want the sans-serif version of your font. Default value
is false and will use the serif version. Also updated the README
accordingly.

NOTE: to pass these options to tuestyle2008.sty, we can only pass a option
without value. This has to do with the fact that tuestyle2008 is not a
BeamerPackage and that we would need to use kvopions or some sort of
other package to facilitate this. As there are onyl two options, the option
is only passed to tuestyle2008 if we need to deviate from the default
behaviour.